requirements.txt
torch
numpy
requests
pyyaml
tqdm
tensorboard
matplotlib 